Fixed GUI issues. Assistant at least runs now
